Rename index.tsx to Input.tsx fixed storybook import.
