+++ src/forms/input/Input.stories.tsx
@@ -11,7 +11,7 @@
 import * as React from "react";
 import {storiesOf} from "@storybook/react";
 import {action} from "@storybook/addon-actions";
-import {Input} from "./index";
+import {Input} from "./Input";
 import {Icon} from "../../icon";
 
 storiesOf("Input", module)
